"Remove content negotiation method and enhance RestControllerMain"
The content negotiation method was removed from the 'MyConfig’ configuration class to simplify the application setup as the default content type is implicitly set to JSON. The 'RestControllerMain' class was updated: 'showAllEmployees' method was refactored for cleaner code by directly returning the list from the service call. Additionally, method 'showEmployeeById' was uncommented and updated to fetch a particular employee, enhancing the API's functionality.