Refactor package names to include 'server'
The commit includes the renaming and refactoring of several java files to include 'server' in their package and class names for clarity and better organization. All associated references have been altered accordingly. It also introduced a new configuration file for the 'consumer' package. This was made to separate 'server' package and 'consumer' package which leads to clearer, more maintainable code structure.