"Add Communication and Employee classes for consumer app"
This commit includes the addition of the Communication.java and Employee.java classes in the consumer application. These are essential for RESTful communication with the server application. The Communication class utilizes Spring's RestTemplate to call to the server API, getting all employee details. It is also equipped with placeholder methods for future functionalities like get, save, and delete for an individual employee. The Employee entity represents the data structure expected from the server. Furthermore, in the server-side Employee class, the redundant @Component annotation is removed. Other changes include modifying the API endpoint in the ControllerMain class on the server side for unambiguous communication. A RestTemplate Bean is also added in consumer's configuration for Autowiring.